requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
mysqlclient==2.2.7
pymysql==1.1.1
pydantic==2.10.6
pydantic[email]==2.10.6
passlib==1.7.4
bcrypt==4.3.0
pandas==2.2.2
python-jose==3.3.0
python-multipart==0.0.9